requirements.txt
dnspython==2.4.2
python-nmap==0.7.1
requests==2.31.0
colorama==0.4.6
python-whois==0.8.0
pyfiglet==1.0.2